Extract crud dialog example
