+++ lib/ui/expense/dialogs/expense_dialogs.dart
@@ -2,3 +2,4 @@ export 'assignee_picker_dialog.dart';
 export 'author_change_dialog.dart';
 export 'expense_settings_dialog.dart';
 export 'receipt_scan_dialog.dart';
+export 'group_invite_dialog.dart';
